# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a code scanner service API.",
        "title": "vibescan API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/advanced-scan/repo": {
            "post": {
                "description": "Enqueues an advanced scan task for a given repository URL and language.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "Trigger a code scan on a repository with advanced scans.",
                "parameters": [
                    {
                        "description": "Request body containing repository URL and language",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.ScanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully enqueued task",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to enqueue scan task",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/progress/{task_id}": {
            "get": {
                "description": "Streams scan progress via Server-Sent Events (SSE)",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Scan"
                ],
                "summary": "Task Progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE Stream",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/reports/owasp/{task_id}": {
            "get": {
                "description": "Generates OWASP report from a specific task result using task ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get OWASP report for a task result.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated OWASP report",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/armur-codescanner_pkg.ReportItem"
                            }
                        }
                    },
                    "404": {
                        "description": "Task result not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch task result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reports/sans/{task_id}": {
            "get": {
                "description": "Generates SANS report from a specific task result using task ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get SANS report for a task result.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated SANS report",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/armur-codescanner_pkg.SANSReportItem"
                            }
                        }
                    },
                    "404": {
                        "description": "Task result not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch task result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/scan/batch": {
            "post": {
                "description": "Enqueues multiple scan tasks at once",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan"
                ],
                "summary": "Batch Scan",
                "parameters": [
                    {
                        "description": "Batch Scan Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.BatchScanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.BatchScanResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scan/file": {
            "post": {
                "description": "Enqueues a scan task for a given file.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "Trigger a code scan on file.",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to be scanned",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successfully enqueued task",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "No file part or no selected file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create temp directory or Failed to save file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/scan/local": {
            "post": {
                "description": "Enqueues a scan task for a given local repository path and language.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "Trigger a code scan on a local repository.",
                "parameters": [
                    {
                        "description": "Request body containing local path and language",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.LocalScanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully enqueued task",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to enqueue scan task",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/scan/repo": {
            "post": {
                "description": "Enqueues a scan task for a given repository URL and language.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scan"
                ],
                "summary": "Trigger a code scan on a repository.",
                "parameters": [
                    {
                        "description": "Request body containing repository URL and language",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.ScanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully enqueued task",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to enqueue scan task",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/scan/{task_id}": {
            "delete": {
                "description": "Cancels an in-progress scan task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan"
                ],
                "summary": "Cancel Scan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "armur-codescanner_pkg.ReportItem": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "integer"
                },
                "confidence": {
                    "type": "string"
                },
                "file": {
                    "type": "string"
                },
                "line": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "owasp": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "suggested_remediation": {
                    "type": "string"
                }
            }
        },
        "armur-codescanner_pkg.SANSReportItem": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "integer"
                },
                "confidence": {
                    "type": "string"
                },
                "cwe": {
                    "type": "string"
                },
                "file": {
                    "type": "string"
                },
                "line": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "suggested_remediation": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.ScanRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string",
                    "example": "go"
                },
                "repository_url": {
                    "type": "string",
                    "example": "https://github.com/Armur-Ai/Armur-Code-Scanner"
                },
                "webhook_secret": {
                    "type": "string",
                    "example": "my-hmac-secret"
                },
                "webhook_url": {
                    "type": "string",
                    "example": "https://hooks.example.com/armur"
                }
            }
        },
...
    }
}